+++ spams/dictLearn/mex/mexTrainDL_Memory.cpp
@@ -46,29 +46,29 @@ template <typename T>
 
       T* prX = reinterpret_cast<T*>(mxGetPr(prhs[0]));
       const mwSize* dimsX=mxGetDimensions(prhs[0]);
-      long n=static_cast<long>(dimsX[0]);
-      long M=static_cast<long>(dimsX[1]);
+      int n=static_cast<int>(dimsX[0]);
+      int M=static_cast<int>(dimsX[1]);
       Matrix<T> X(prX,n,M);
 
-      long NUM_THREADS = getScalarStructDef<long>(prhs[1],"numThreads",-1);
+      int NUM_THREADS = getScalarStructDef<int>(prhs[1],"numThreads",-1);
 #ifdef _OPENMP
       NUM_THREADS = NUM_THREADS == -1 ? omp_get_num_procs() : NUM_THREADS;
 #else
       NUM_THREADS=1;
 #endif 
-      long batch_size = getScalarStructDef<long>(prhs[1],"batchsize",
+      int batch_size = getScalarStructDef<int>(prhs[1],"batchsize",
             256*(NUM_THREADS+1));
       mxArray* pr_D = mxGetField(prhs[1],0,"D");
       Trainer<T>* trainer;
 
       if (!pr_D) {
-         long K = getScalarStruct<long>(prhs[1],"K");
+         int K = getScalarStruct<int>(prhs[1],"K");
          trainer = new Trainer<T>(K,batch_size,NUM_THREADS);
       } else {
          T* prD = reinterpret_cast<T*>(mxGetPr(pr_D));
          const mwSize* dimsD=mxGetDimensions(pr_D);
-         long nD=static_cast<long>(dimsD[0]);
-         long K=static_cast<long>(dimsD[1]);
+         int nD=static_cast<int>(dimsD[0]);
+         int K=static_cast<int>(dimsD[1]);
          if (n != nD) mexErrMsgTxt("sizes of D are not consistent");
          Matrix<T> D1(prD,n,K);
          trainer = new Trainer<T>(D1,batch_size,NUM_THREADS);
@@ -76,23 +76,23 @@ template <typename T>
 
       ParamDictLearn<T> param;
       param.lambda = getScalarStruct<T>(prhs[1],"lambda");
-      param.iter=getScalarStruct<long>(prhs[1],"iter");
-      param.mode = (constraint_type)getScalarStructDef<long>(prhs[1],"mode",PENALTY);
+      param.iter=getScalarStruct<int>(prhs[1],"iter");
+      param.mode = (constraint_type)getScalarStructDef<int>(prhs[1],"mode",PENALTY);
       if (param.mode != PENALTY && param.mode != L2ERROR) 
          mexErrMsgTxt("param.mode is not compatible with the offline setting");
       param.posD = getScalarStructDef<bool>(prhs[1],"posD",false);
-      param.modeD=(constraint_type_D)(getScalarStructDef<long>(prhs[1],"modeD",0));
+      param.modeD=(constraint_type_D)(getScalarStructDef<int>(prhs[1],"modeD",0));
       param.whiten = getScalarStructDef<bool>(prhs[1],"whiten",false);
-      param.modeParam = static_cast<mode_compute>(getScalarStructDef<long>(prhs[1],"modeParam",0));
+      param.modeParam = static_cast<mode_compute>(getScalarStructDef<int>(prhs[1],"modeParam",0));
       param.clean = getScalarStructDef<bool>(prhs[1],"clean",true);
       param.gamma1 = getScalarStructDef<T>(prhs[1],"gamma1",0);
       param.gamma2 = getScalarStructDef<T>(prhs[1],"gamma2",0);
       param.rho = getScalarStructDef<T>(prhs[1],"rho",T(1.0));
-      param.iter_updateD = getScalarStructDef<long>(prhs[1],"iter_udpateD",1);
+      param.iter_updateD = getScalarStructDef<int>(prhs[1],"iter_udpateD",1);
       trainer->trainOffline(X,param);
       Matrix<T> D;
       trainer->getD(D);
-      long K  = D.n();
+      int K  = D.n();
       plhs[0] = createMatrix<T>(n,K);
       T* prD2 = reinterpret_cast<T*>(mxGetPr(plhs[0]));
       Matrix<T> D2(prD2,n,K);
